+++ Changelog.md
@@ -5,6 +5,7 @@
 1. Refactored input system to its own class
     1. Externalized input delimiter configuration
 2. Externalized coupon data source - no need to change code for adding / updating coupons
-3. Added tests:
+3. New, optimized ETA calculation
+4. Added tests:
     1. Input validations
-    2. ETA calculations
+    2. ETA calculations for both simple and optimized implementations